backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "django-cors-headers>=4.9.0",
    "django>=6.0.3",
    "djangorestframework>=3.16.1",
    "djoser>=2.3.3",
]
